backend/pyproject.toml
[project]
name = "ournote"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.124.4",
    "psycopg[binary]>=3.3.2",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.45",
]

[tool.ruff]
src = ["app"]

[tool.ty.environment]
python = ".venv"
root = ["./"]

[tool.ty.rules]
invalid-argument-type = "ignore"  # Pydantic model constructors
invalid-key = "ignore"            # TypedDict/Pydantic interop
invalid-parameter-default = "ignore"  # Field defaults

[dependency-groups]
dev = [
    "ruff>=0.14.10",
    "ty>=0.0.8",
]
